+++ src/pages/Favorites/index.js
@@ -1,21 +1,14 @@
-import React, { useState } from "react";
-import {
-    View,
-    Text,
-    TouchableOpacity,
-    Image,
-    FlatList,
-    PanResponder,
-    Animated,
-} from "react-native";
+import React, { Component } from "react";
+import DraggableFlatList from "react-native-draggable-flatlist";
 import Swipeable from "react-native-swipeable";
+import { View, TouchableOpacity, Text, Image } from "react-native";
 import { connect } from "react-redux";
-
 import { Feather } from "@expo/vector-icons";
-import { MaterialCommunityIcons } from "@expo/vector-icons";
 
-import styles from "./styles";
+import { MaterialCommunityIcons } from "@expo/vector-icons";
+import globalStyles from "../../globalStyle/globalStyles";
 import profilePic from "../../assets/defaultUserImage.png";
+import styles from "./styles";
 
 import Header from "../../components/Header";
 
@@ -30,150 +23,62 @@ const btns = [
     </TouchableOpacity>,
 ];
 
-class Favorites extends React.Component {
-    state = {
-        dragging: false,
-        draggingIdx: -1,
-        // data: [
-        //     { id: 1, name: "Talita Cypriano" },
-        //     { id: 2, name: "André Leme" },
-        //     { id: 3, name: "Emílio Rodrigues" },
-        //     { id: 4, name: "Outro Professor" },
-        //     { id: 5, name: "Outra Professora" },
-        //     { id: 6, name: "Outra Professora" },
-        // ]
-    };
-
-    point = new Animated.ValueXY();
-    scrollOffset = 0;
-    flatListTopOffset = 0;
-    rowHeight = 0;
-    currentIdx = -1;
-
-    constructor(props) {
-        super(props);
-
-        this._panResponder = PanResponder.create({
-            // Ask to be the responder:
-            onStartShouldSetPanResponder: (evt, gestureState) => true,
-            onStartShouldSetPanResponderCapture: (evt, gestureState) => true,
-            onMoveShouldSetPanResponder: (evt, gestureState) => true,
-            onMoveShouldSetPanResponderCapture: (evt, gestureState) => true,
-
-            onPanResponderGrant: (evt, gestureState) => {
-                console.log(gestureState.y0);
-                this.currentIdx = this.yToIndex(gestureState.y0);
-                this.setState({ dragging: true, draggingIdx: this.currentIdx });
-                // The gesture has started. Show visual feedback so the user knows
-                // what is happening!
-                // gestureState.d{x,y} will be set to zero now
-            },
-            onPanResponderMove: (evt, gestureState) => {
-                Animated.event([{ y: this.point.y }])({
-                    y: gestureState.moveY,
-                });
-                // The most recent move distance is gestureState.move{X,Y}
-                // The accumulated gesture distance since becoming responder is
-                // gestureState.d{x,y}
-            },
-            onPanResponderTerminationRequest: (evt, gestureState) => false,
-            onPanResponderRelease: (evt, gestureState) => {
-                // The user has released all touches while this view is the
-                // responder. This typically means a gesture has succeeded
-                this.reset();
-            },
-            onPanResponderTerminate: (evt, gestureState) => {
-                // Another component has become the responder, so this gesture
-                // should be cancelled
-                this.reset();
-            },
-            onShouldBlockNativeResponder: (evt, gestureState) => {
-                // Returns whether this component should block native components from becoming the JS
-                // responder. Returns true by default. Is currently only supported on android.
-                return true;
-            },
-        });
-    }
-
-    yToIndex = (y) => {
-        Math.floor(
-            (this.scrollOffset + y - this.flatListTopOffset) / this.rowHeight
-        );
-    };
-
-    reset = () => {
-        this.setState({ dragging: false });
+class Favorites extends Component {
+    changeOrder = (newState) => {
+        this.props.changeOrder(newState.data);
     };
 
     deleteProf = (id) => {
         this.props.deleteProf(id);
-        // setProfList((currentList) => {
-        //     return currentList.filter((el) => {
-        //         return el.id !== id;
-        //     });
-        // });
     };
-    render() {
-        const { favorProfs } = this.props;
 
-        const { dragging, draggingIdx } = this.state;
-
-        const renderItem = ({ item: prof }) => (
+    renderItem = ({ item, index, drag, isActive }) => {
+        return (
             <Swipeable
                 leftButtons={btns}
-                onLeftActionRelease={() => deleteProf(prof.id)}
+                onLeftActionRelease={() => deleteProf(item.id)}
             >
-                <View
-                    style={styles.listContainer}
-                    {...this._panResponder.panHandlers}
-                    onLayout={(e) => {
-                        this.rowHeight = e.nativeEvent.layout.height;
+                <TouchableOpacity
+                    style={{
+                        ...styles.itemContainer,
                     }}
+                    onPressIn={drag}
                 >
-                    <MaterialCommunityIcons name="drag" size={30} />
-                    <Image source={profilePic} style={styles.profilePic} />
-                    <Text style={styles.profName}>{prof.name}</Text>
-                </View>
+                    <View
+                        style={{
+                            ...styles.listContainer,
+                            elevation: isActive ? 3 : 1,
+                        }}
+                    >
+                        <MaterialCommunityIcons name="drag" size={40} />
+                        <Image source={profilePic} style={styles.profilePic} />
+                        <Text style={styles.profName}>{item.name}</Text>
+                    </View>
+                </TouchableOpacity>
             </Swipeable>
         );
+    };
 
+    render() {
+        const { favorProfs } = this.props;
         return (
-            <>
+            <View style={globalStyles.container}>
                 <Header />
 
                 <View style={styles.titleContainer}>
                     <Text style={styles.title}>Salvos</Text>
                 </View>
-
-                {dragging && (
-                    <Animated.View
-                        style={{
-                            position: "absolute",
-                            zIndex: 2,
-                            backgroundColor: "red",
-                            width: "100%",
-                            top: this.point.getLayout().top,
-                        }}
-                    >
-                        {renderItem({ prof: draggingIdx })}
-                    </Animated.View>
-                )}
-
-                {favorProfs.length ? (
-                    <FlatList
-                        scrollEnabled={!dragging}
-                        onScroll={(e) => {
-                            this.scrollOffset = e.nativeEvent.contentOffset.y;
-                        }}
-                        onLayout={(e) => {
-                            this.flatListTopOffset = e.nativeEvent.layout.y;
-                        }}
-                        style={styles.favList}
-                        data={favorProfs}
-                        showsVerticalScrollIndicator={false}
-                        keyExtractor={(favorProfs) => String(favorProfs.id)}
-                        renderItem={renderItem}
-                    />
+                {favorProfs ? (
+                    <View style={styles.draggableContainer}>
+                        <DraggableFlatList
+                            data={favorProfs}
+                            renderItem={this.renderItem}
+                            keyExtractor={(item, index) =>
+                                `draggable-item-${item.id}`
+                            }
+                            onDragEnd={({ data }) => this.changeOrder({ data })}
+                        />
+                    </View>
                 ) : (
                     <View style={styles.msgContainer}>
                         <Text style={styles.msg}>
@@ -181,7 +86,7 @@ class Favorites extends React.Component {
                         </Text>
                     </View>
                 )}
-            </>
+            </View>
         );
     }
 }
@@ -197,6 +102,9 @@ const mapDispatchToProps = (dispatch) => {
     return {
         deleteProf: (id) => {
             dispatch({ type: "DELETE_PROF", id: id });
+        },
+        changeOrder: (newState) => {
+            dispatch({ type: "CHANGE_ORDER", newState: newState });
         },
     };
 };